display exports for local packages
Display exports for local packages in ‘meteor show’. (We will eventually display exports for non-local packages as well)

- method to aggregate exports for a package in packageSource (exports are per-architecture).

- get this data from packageSource in PackageQuery for ‘meteor show’. Don’t store it in the local catalog — while it is not a particularly expensive operation, it is slightly more than a simple lookup.
We really care about minimizing any sort of computation when we are initializing packages, since we
want the tool to be fast.

- display the data in ‘meteor show’. It makes sense to line wrap this with the ‘Exports:’ label as a
bulletPoint (just look at the test to see an example where this improves user experience). Since we
are doing that, we might as well use that bulletPoint functionality on the other labels as well.

There is also a test. Run ‘meteor self-test show’ to test, or run ‘meteor show’ on a local package
with exports.